+++ mitmproxy/proxy/layers/http/_http1.py
@@ -11,6 +11,7 @@ from h11._receivebuffer import ReceiveBuffer
 from ...context import Context
 from ._base import format_error
 from ._base import HttpConnection
+from ._events import ErrorCode
 from ._events import HttpEvent
 from ._events import RequestData
 from ._events import RequestEndOfMessage
@@ -98,7 +99,9 @@ class Http1Connection(HttpConnection, metaclass=abc.ABCMeta):
                 yield commands.CloseConnection(self.conn)
                 yield ReceiveHttp(
                     self.ReceiveProtocolError(
-                        self.stream_id, f"HTTP/1 protocol error: {e}"
+                        self.stream_id,
+                        f"HTTP/1 protocol error: {e}",
+                        code=self.ReceiveProtocolError.code,
                     )
                 )
                 return
@@ -137,7 +140,7 @@ class Http1Connection(HttpConnection, metaclass=abc.ABCMeta):
                 self.ReceiveProtocolError(
                     self.stream_id,
                     f"Client disconnected.",
-                    code=status_codes.CLIENT_CLOSED_REQUEST,
+                    code=ErrorCode.CLIENT_DISCONNECTED,
                 )
             )
         else:  # pragma: no cover
@@ -266,9 +269,10 @@ class Http1Server(Http1Connection):
         elif isinstance(event, ResponseProtocolError):
             if not (self.conn.state & ConnectionState.CAN_WRITE):
                 return
-            if not self.response and event.code != status_codes.NO_RESPONSE:
+            status = event.code.http_status_code()
+            if not self.response and status is not None:
                 yield commands.SendData(
-                    self.conn, make_error_response(event.code, event.message)
+                    self.conn, make_error_response(status, event.message)
                 )
             yield commands.CloseConnection(self.conn)
         else:
@@ -294,7 +298,9 @@ class Http1Server(Http1Connection):
                             RequestHeaders(self.stream_id, self.request, False)
                         )
                         yield ReceiveHttp(
-                            RequestProtocolError(self.stream_id, str(e), 400)
+                            RequestProtocolError(
+                                self.stream_id, str(e), ErrorCode.GENERIC_CLIENT_ERROR
+                            )
                         )
                     else:
                         yield commands.Log(
@@ -412,7 +418,9 @@ class Http1Client(Http1Connection):
                     yield commands.CloseConnection(self.conn)
                     yield ReceiveHttp(
                         ResponseProtocolError(
-                            self.stream_id, f"Cannot parse HTTP response: {e}"
+                            self.stream_id,
+                            f"Cannot parse HTTP response: {e}",
+                            ErrorCode.GENERIC_SERVER_ERROR,
                         )
                     )
                     return
@@ -434,6 +442,7 @@ class Http1Client(Http1Connection):
                         ResponseProtocolError(
                             self.stream_id,
                             f"unexpected server response: {bytes(self.buf)!r}",
+                            ErrorCode.GENERIC_SERVER_ERROR,
                         )
                     )
                 else:
@@ -442,7 +451,9 @@ class Http1Client(Http1Connection):
                     # https://tools.ietf.org/html/rfc7231#section-6.5.11
                     yield ReceiveHttp(
                         ResponseProtocolError(
-                            self.stream_id, "server closed connection"
+                            self.stream_id,
+                            "server closed connection",
+                            ErrorCode.GENERIC_SERVER_ERROR,
                         )
                     )
             else: